Updated Forwarder and all tests Passing
Meta transaction forwarder is tested and complete. Further patches could include a limit on users calls and on-chain subscription updates and verification. In order for this contract to be compatible with the VennityCollection contract, the Forwarder.sol address must be hardcoded into the admin list after deployed because currently only the deployer can mint and transfer.